+++ resources/V1_2025.03.12__update_feed_unique_key.sql
@@ -1,0 +1,7 @@
+drop index if exists feed_items_dao_proposal_uindex;
+
+drop index if exists feed_items_unique_index;
+
+create unique index if not exists feed_items_unique_index
+    on feed_items (dao_id, proposal_id, type, action)
+    where type != 'delegate';